requirements.txt
ctransformers==0.2.27
sentence-transformers==3.0.0
pinecone-client
langchain==0.2.1
flask
pypdf
python-dotenv
-e .